Added `show('marker_type_<rank>')`
This outputs the name of the *type* of the marker, e.g. "Chapter", or "Scene".
Performance-wise, this should be nearly a freebie as it's just fetching one more field.

Perhaps in the future we can also output `marker_id_<rank>`.
For now I did not include it because there's not much use for it.
If archives ever start outputting marker IDs, it'll become useful for
linking to the relevant part of the archive :D