Check for divide by zero in thumbnail creation
Problem:
It's possible to divide by zero if an invalid image is fetched by getimagesize(). This crashes the create_thumbnail() function.

Solution:
Check to see if divisor is zero and exit if so.

Testing:
Ran this fix on my website where I had some invalid comic pages in my database and the create_thumbnail() function was able to pass over the invalid images.